+++ pebbling_apps/bookmarks/views/api.py
@@ -1,0 +1,27 @@
+"""API views for bookmarks."""
+
+from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
+from django.views.decorators.http import require_GET
+
+from pebbling_apps.unfurl.unfurl import UnfurlMetadata
+
+
+@login_required
+@require_GET
+def fetch_unfurl_metadata(request):
+    """Fetch and return UnfurlMetadata for a given URL.
+    Primarily in support of pc-bookmark-form
+    """
+    url = request.GET.get("href")
+    if not url:
+        return JsonResponse({"error": "Missing href parameter"}, status=400)
+    try:
+        metadata = UnfurlMetadata(url=url)
+        metadata.unfurl()
+        out = metadata.to_dict(omit_html=True)
+        out["title"] = metadata.title
+        out["description"] = metadata.description
+        return JsonResponse(out)
+    except Exception as e:
+        return JsonResponse({"error": str(e)}, status=500)
