+++ pebbling_apps/bookmarks/views/bookmarks.py
@@ -1,0 +1,154 @@
+"""CRUD views for bookmarks."""
+
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.urls import reverse_lazy
+from django.shortcuts import render, redirect
+from django.views.generic import CreateView, UpdateView, DeleteView
+from django.core.exceptions import PermissionDenied
+from pebbling_apps.unfurl.unfurl import UnfurlMetadata
+
+from ..models import Bookmark
+from ..forms import BookmarkForm
+from .base import BookmarkQueryListView
+
+
+class BookmarkListView(BookmarkQueryListView):
+    """List View: Show all bookmarks for the logged-in user."""
+
+    model = Bookmark
+    template_name = "bookmarks/bookmark_list.html"
+    context_object_name = "bookmarks"
+
+
+class BookmarkCreateView(LoginRequiredMixin, CreateView):
+    """Create View: Add a new bookmark."""
+
+    model = Bookmark
+    template_name = "bookmarks/bookmark_form.html"
+    form_class = BookmarkForm
+    success_url = reverse_lazy("bookmarks:list")
+
+    def get_initial(self):
+        """Pre-populate form with query parameters or existing bookmark data."""
+        initial = super().get_initial()
+        url = self.request.GET.get("url", "")
+
+        if url and self.request.user.is_authenticated:
+            self.existing_bookmark = Bookmark.objects.filter(
+                owner=self.request.user,
+                unique_hash=Bookmark.objects.generate_unique_hash_for_url(url),
+            ).first()
+
+            if self.existing_bookmark:
+                initial.update(
+                    {
+                        "url": self.existing_bookmark.url,
+                        "title": self.existing_bookmark.title,
+                        "description": self.existing_bookmark.description,
+                        "tags": " ".join(
+                            tag.name for tag in self.existing_bookmark.tags.all()
+                        ),
+                        "unfurl_metadata": self.existing_bookmark.unfurl_metadata,
+                    }
+                )
+                return initial
+
+            # No existing bookmark - try to fetch metadata
+            try:
+                unfurl_metadata = UnfurlMetadata(url=url)
+                unfurl_metadata.unfurl()  # Fetch and parse metadata
+
+                initial.update(
+                    {
+                        "url": url,
+                        "title": self.request.GET.get("title", unfurl_metadata.title),
+                        "description": self.request.GET.get(
+                            "description",
+                            unfurl_metadata.description,
+                        ),
+                        "tags": self.request.GET.get("tags", ""),
+                        "unfurl_metadata": unfurl_metadata,
+                    }
+                )
+            except Exception as e:
+                # If metadata fetch fails, just use query parameters
+                initial.update(
+                    {
+                        "url": url,
+                        "title": self.request.GET.get("title", ""),
+                        "description": self.request.GET.get("description", ""),
+                        "tags": self.request.GET.get("tags", ""),
+                    }
+                )
+
+        return initial
+
+    def get_context_data(self, **kwargs):
+        """Add existing bookmark info and layout mode to context."""
+        context = super().get_context_data(**kwargs)
+        if hasattr(self, "existing_bookmark"):
+            context["existing_bookmark"] = self.existing_bookmark
+        # Add minimal_layout flag if popup parameter is present
+        context["minimal_layout"] = "popup" in self.request.GET
+        return context
+
+    def get_form_kwargs(self):
+        """Pass the current user to the form."""
+        kwargs = super().get_form_kwargs()
+        kwargs["user"] = self.request.user
+        return kwargs
+
+    def form_valid(self, form):
+        """Handle successful form submission with custom redirect logic."""
+        form.instance.owner = self.request.user
+        self.object = form.save()
+
+        # Check for next parameter
+        next_param = self.request.GET.get("next")
+        if next_param == "close":
+            return render(
+                self.request,
+                "bookmarks/bookmark_create_close.html",
+                {"bookmark": self.object},
+            )
+        elif next_param == "profile":
+            return redirect("profiles:view", username=self.request.user.username)
+        elif next_param == "same":
+            return redirect(self.object.url)
+
+        return super().form_valid(form)
+
+
+class BookmarkUpdateView(LoginRequiredMixin, UpdateView):
+    """Update View: Edit an existing bookmark."""
+
+    model = Bookmark
+    form_class = BookmarkForm
+    template_name = "bookmarks/bookmark_form.html"
+    success_url = reverse_lazy("bookmarks:list")
+
+    def get_form_kwargs(self):
+        """Pass the current user to the form."""
+        kwargs = super().get_form_kwargs()
+        kwargs["user"] = self.request.user
+        return kwargs
+
+    def get_object(self, queryset=None):
+        bookmark = super().get_object(queryset)
+        if bookmark.owner != self.request.user:
+            raise PermissionDenied  # Prevent unauthorized editing
+        return bookmark
+
+
+class BookmarkDeleteView(LoginRequiredMixin, DeleteView):
+    """Delete View: Remove a bookmark."""
+
+    model = Bookmark
+    template_name = "bookmarks/bookmark_confirm_delete.html"
+    success_url = reverse_lazy("bookmarks:list")
+
+    def get_object(self, queryset=None):
+        bookmark = super().get_object(queryset)
+        if bookmark.owner != self.request.user:
+            raise PermissionDenied  # Prevent unauthorized deletion
+        return bookmark
